requirements.txt
Flask>=3.0.0,<4.0.0
Werkzeug>=3.0.0,<4.0.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
Pillow>=10.4.0
numpy>=2.3.0
pandas>=2.2.3
gunicorn>=21.2.0
flask-cors>=4.0.0
opencv-python-headless>=4.10.0.84
PyMuPDF>=1.24.10
easyocr>=1.7.2
